task: Add proxy url settings and fix HMR error
Add Lando configuration settings to allow UI to be accessed by `spark-ui.local` as well as instructions for making it available in the local dev environment.

Changed the Lando project name to distinguish UI and API naming.

Add a `port` setting to the `vite.hmr` configuration to fix a websocket error in the browser console.