task: Add Ray configuration to Site directory
Add the Ray configuration file (`ray.php`) to the project's `site` directory. This initially only existed in the root directory and worked fine if I was starting the Lando environment from there. This fails if I startup the Lando project at the `site` level in isolation from the other project components.